# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MS Elevate LEAPS Tracker API
  version: 1.0.0
  description: |
    Complete API documentation for the Microsoft Elevate Indonesia LEAPS Tracker platform.
    
    The LEAPS framework consists of 5 stages:
    - **Learn**: Complete Microsoft Learn courses (20 points)
    - **Explore**: Apply AI in classroom with evidence (50 points)  
    - **Amplify**: Train peers/students (2 points/peer, 1 point/student)
    - **Present**: LinkedIn post with screenshot (20 points)
    - **Shine**: Innovation ideas for recognition
    
    ## Authentication
    All authenticated endpoints require a valid Clerk JWT token in the Authorization header.
    
    ## Rate Limits
    - File uploads: 10MB max per file
    - Amplify submissions: 50 peers, 200 students per 7-day rolling window
    - Bulk operations: Maximum 50 items per request
    
  contact:
    name: MS Elevate Team
    url: https://leaps.mereka.org
servers:
  - url: https://leaps.mereka.org/api
    description: Production server
  - url: https://staging-leaps.mereka.org/api 
    description: Staging server
security:
  - ClerkJWT: []
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication JWT token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (for Zod errors)
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
        handle:
          type: string
        name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
        school:
          type: string
        cohort:
          type: string
        role:
          type: string
          enum: [participant, reviewer, admin, superadmin]
        profile_visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - handle
        - name
        - email
    Activity:
      type: object
      properties:
        code:
          type: string
          enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        name:
          type: string
        description:
          type: string
        default_points:
          type: integer
      required:
        - code
        - name
        - default_points
    Submission:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        activity_code:
          type: string
          enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        payload:
          type: object
          description: Activity-specific data
        attachments:
          type: array
          items:
            type: string
        reviewer_id:
          type: string
          nullable: true
        review_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - activity_code
        - status
        - payload
...
paths:
...
  /admin/submissions:
    get:
      summary: Get submissions queue for review
      tags: [Admin]
      security:
        - ClerkJWT: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, ALL]
            default: PENDING
        - name: activity
          in: query
          schema:
            type: string
            enum: [LEARN, EXPLORE, AMPLIFY, PRESENT, SHINE]
        - name: userId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: sortBy
          in: query
          schema:
            type: string
            default: created_at
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated submissions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  submissions:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Submission'
                        - type: object
                          properties:
                            user:
                              $ref: '#/components/schemas/User'
                            activity:
                              $ref: '#/components/schemas/Activity'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Review a submission (approve/reject)
      tags: [Admin]
      security:
        - ClerkJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionId:
                  type: string
                action:
                  type: string
                  enum: [approve, reject]
                reviewNote:
                  type: string
                pointAdjustment:
                  type: integer
                  description: Override points (within ±20% of base)
              required:
                - submissionId
                - action
      responses:
        '200':
          description: Submission reviewed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  submission:
                    $ref: '#/components/schemas/Submission'
                  message:
                    type: string
        '400':
          description: Invalid request or submission already reviewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Bulk review submissions
      tags: [Admin]
      security:
        - ClerkJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submissionIds:
                  type: array
                  items:
                    type: string
                  maxItems: 50
                action:
                  type: string
                  enum: [approve, reject]
                reviewNote:
                  type: string
              required:
                - submissionIds
                - action
      responses:
        '200':
          description: Bulk operation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                  message:
                    type: string
        '400':
          description: Invalid request or too many submissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/kajabi/reprocess:
    post:
      summary: Reprocess failed Kajabi webhook events
      tags: [Admin]
      security:
        - ClerkJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventIds:
                  type: array
                  items:
                    type: string
                  maxItems: 50
              required:
                - eventIds
      responses:
        '200':
          description: Events reprocessed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        eventId:
                          type: string
                        error:
                          type: string
...